feat: Mock Socket IO
