+++ src/hooks/useSocketIO.tsx
@@ -1,0 +1,4 @@
+import { useContext } from 'react'
+import { SocketIOContext } from 'src/context/SocketIOContext'
+
+export const useSocketIO = () => useContext(SocketIOContext)
