+++ src/pages/login/index.tsx
@@ -130,7 +130,8 @@ const LoginPage = () => {
                     <CustomTextField
                       fullWidth
                       autoFocus
-                      label='Email'
+                      label='E-mail'
+                      type='email'
                       value={value}
                       onBlur={onBlur}
                       onChange={onChange}
