+++ src/pages/revendas/criar-revenda/index.tsx
@@ -177,6 +177,7 @@ const CreateResale = () => {
                   <CustomTextField
                     fullWidth
                     label='E-mail'
+                    type='email'
                     value={value}
                     onBlur={onBlur}
                     onChange={onChange}
