+++ src/services/socketIO/index.ts
@@ -1,0 +1,3 @@
+import io from 'socket.io-client'
+
+export const socketIO = io
